+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Release History
 
 - errno to Exception mapping became a fixed part of the API (and unit tested)
 - timerfd is now CLOCK_MONOTONIC by default
+- Added Simplier TimerVal to replace Timerspec for timerfd
 
 **Bug Fixes**